Remove some other unnecessary code
