--- Patched Sur - For Catalina/Assets.xcassets/QuestionMark.imageset/Contents.json
@@ -1,26 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "QuestionMark.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "QuestionMark@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "QuestionMark@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "template-rendering-intent" : "template"
-  }
-}
