--- Patched Sur - For Catalina/Assets.xcassets/Warn2.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Warn2.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "Warn2@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "Warn2@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
