+++ Patched Sur - For Catalina/Views/DownloadView.swift
@@ -66,3 +66,20 @@ struct DownloadKextsView: View {
         }
     }
 }
+
+struct ProgressBar: View {
+    @Binding var value: CGFloat
+    var length: CGFloat = 285
+    
+    var body: some View {
+        ZStack(alignment: .leading) {
+            Rectangle().frame(minWidth: length)
+                .opacity(0.3)
+                .foregroundColor(Color("Accent").opacity(0.9))
+            
+            Rectangle().frame(width: min(value*length, length))
+                .foregroundColor(Color("Accent"))
+                .animation(.linear)
+        }.cornerRadius(20)
+    }
+}
